requirements.txt
langchain
openai
streamlit
python-dotenv
huggingface_hub
